Fold telemetry tests into each block's index.test.ts
Move the AuthBasic and Logger telemetry-registration assertions out of a
dedicated telemetry.test.ts and into each package's existing index.test.ts,
removing the now-empty dedicated files. The runner globs all *.test.js, so
coverage is unchanged; this keeps both blocks consistent with the repo's
prevailing one-test-file-per-package layout.
