+++ packages/bb-auth-basic/src/index.test.ts
@@ -8,12 +8,18 @@
  * `hasAuthError` instead of string-matching the human-facing message.
  */
 
-import { describe, test } from 'node:test';
+import { describe, test, beforeEach } from 'node:test';
 import assert from 'node:assert';
+import { readFileSync } from 'node:fs';
+import { dirname, join } from 'node:path';
+import { fileURLToPath } from 'node:url';
 import type { BlocksContext } from '@aws-blocks/core';
 import { Scope, hasAuthError } from '@aws-blocks/core';
 import type { AuthStateApi } from '@aws-blocks/auth-common';
 import { AuthBasic, AuthBasicErrors, type AuthBasicOptions } from './index.js';
+import { BB_NAME, BB_VERSION } from './version.js';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 function ctx(origin?: string): BlocksContext {
 	const headers = new Headers();
@@ -72,5 +78,46 @@ describe('AuthBasic setAuthState errorName', () => {
 		assert.strictEqual(next.state, 'signedOut');
 		assert.ok(next.error);
 		assert.strictEqual(next.errorName, undefined);
+	});
+});
+
+/**
+ * Telemetry-registration tests for AuthBasic.
+ *
+ * `Scope.getRegisteredBlocks()` only names a block whose `bbName` is in
+ * OFFICIAL_BB_NAMES, and that set is generated from the umbrella's
+ * `aws-blocks.vendorize` map. These tests pin the three coupled artifacts to
+ * each other: the block's generated BB_NAME, its vendorize entry, and the
+ * generated name set. A block that omits `bbMeta` still constructs fine and
+ * every other test still passes, so that gap is only visible here.
+ */
+describe('AuthBasic telemetry registration', () => {
+	beforeEach(() => {
+		Scope._resetRegistry();
+	});
+
+	test('BB_NAME is the name the vendorize map and OFFICIAL_BB_NAMES carry', () => {
+		assert.strictEqual(BB_NAME, 'AuthBasic');
+	});
+
+	test('BB_VERSION tracks the package version', () => {
+		const pkg = JSON.parse(readFileSync(join(__dirname, '..', 'package.json'), 'utf-8'));
+		assert.strictEqual(BB_VERSION, pkg.version);
+	});
+
+	test('an instance carries bbName and bbVersion', () => {
+		const auth = makeAuth();
+		assert.strictEqual(auth.bbName, BB_NAME);
+		assert.strictEqual(auth.bbVersion, BB_VERSION);
+	});
+
+	test('registers as an official block, so telemetry is allowed to name it', () => {
+		makeAuth();
+		const { blocks, customBlocksCount } = Scope.getRegisteredBlocks();
+		assert.deepStrictEqual(
+			blocks.filter(b => b.name === BB_NAME),
+			[{ name: BB_NAME, version: BB_VERSION }],
+		);
+		assert.strictEqual(customBlocksCount, 0, 'must not be filtered out as an unnamed custom block');
 	});
 });
